fix IGMIdx type in operators

+++ RedshiftLibrary/RedshiftLibrary/operator/templatefitting.h
@@ -34,8 +34,8 @@ public:
                                               Int32 opt_dustFitting=0,
                                               CPriorHelper::TPriorZEList logpriorze=CPriorHelper::TPriorZEList(),
                                               Bool keepigmism = false,
-                                              Float64 FitEbmvCoeff=-1,
-                                              Float64 FitMeiksinIdx=-1);
+                                              Float64 FitEbmvCoeff=-1.,
+                                              Int32 FitMeiksinIdx=-1);
 
     const COperatorResult* ExportChi2versusAZ( const CSpectrum& spectrum, const CTemplate& tpl,
                                     const TFloat64Range& lambdaRange, const TFloat64List& redshifts,
@@ -58,7 +58,7 @@ private:
                   Float64& fittingMtM,
                   Float64 &fittingLogprior,
                   Float64 &fittingEbmvCoeff,
-                  Float64 &fittingMeiksinIdx,
+                  Int32 &fittingMeiksinIdx,
                   EStatus& status,
                   std::vector<TFloat64List>& ChiSquareInterm,
                   std::vector<TFloat64List>& IsmCalzettiCoeffInterm,
